Extract sample contract and brief paragraphs for each example
