+++ documentation/public/user-guide/08-examples/01-list-functions-in-contract/examples/test-list-functions.test.mts
@@ -1,31 +1,9 @@
 import assert from "node:assert";
-import { buildCompilationUnit } from "../../common/compilation-builder.mjs";
 import { listFunctionsInContract } from "./list-functions-in-contract.mjs";
+import buildSampleCompilationUnit from "../../common/sample-contract.mjs";
 
 test("list functions by contract name", async () => {
-  const CONTRACT_VFS = new Map<string, string>([
-    [
-      "contract.sol",
-      `
-contract Counter {
-  uint _count;
-  constructor(uint initialCount) {
-    _count = initialCount;
-  }
-  function count() public view returns (uint) {
-    return _count;
-  }
-  function increment(uint delta) public returns (uint) {
-    require(delta > 0, "Delta must be positive");
-    _count += delta;
-    return _count;
-  }
-}
-      `,
-    ],
-  ]);
-
-  const unit = await buildCompilationUnit(CONTRACT_VFS, "0.8.0", "contract.sol");
+  const unit = await buildSampleCompilationUnit();
 
   const functions = listFunctionsInContract(unit, "Counter");
   assert(functions);
