+++ documentation/public/user-guide/08-examples/02-find-usages/examples/test-find-usages.test.mts
@@ -1,31 +1,9 @@
 import assert from "node:assert";
-import { buildCompilationUnit } from "../../common/compilation-builder.mjs";
 import { findUsages } from "./find-usages.mjs";
+import buildSampleCompilationUnit from "../../common/sample-contract.mjs";
 
 test("find usages", async () => {
-  const CONTRACT_VFS = new Map<string, string>([
-    [
-      "contract.sol",
-      `
-contract Counter {
-  uint _count;
-  constructor(uint initialCount) {
-    _count = initialCount;
-  }
-  function count() public view returns (uint) {
-    return _count;
-  }
-  function increment(uint delta) public returns (uint) {
-    require(delta > 0, "Delta must be positive");
-    _count += delta;
-    return _count;
-  }
-}
-      `,
-    ],
-  ]);
-
-  const unit = await buildCompilationUnit(CONTRACT_VFS, "0.8.0", "contract.sol");
+  const unit = await buildSampleCompilationUnit();
 
   const usages = findUsages(unit, "contract.sol", 2, 10); // the _count state variable definition
   assert(typeof usages !== "string");
